Moved Chatterbox launch code into a module (transport.so).
Having multiple branches to keep track of the differences between
launching stage simulations versus chatterbox runs was annoying.
Moreover, there was no good argument that 'the code running is the
same in simulation and reality'. No compilation changes are now
required for either target and the source is the same; an
InitCB function does for the Chatterbox what Init does for stage.